chore: AgeModal Controller 분리
